Move automation trace support to its own module
