+++ tests/components/automation/test_websocket_api.py
@@ -250,19 +250,19 @@ async def test_automation_trace_overflow(hass, hass_ws_client):
     assert len(response["result"]["sun"]) == 1
 
     # Trigger "moon" automation enough times to overflow the number of stored traces
-    for _ in range(automation.STORED_TRACES):
+    for _ in range(automation.trace.STORED_TRACES):
         hass.bus.async_fire("test_event2")
         await hass.async_block_till_done()
 
     await client.send_json({"id": next_id(), "type": "automation/trace/list"})
     response = await client.receive_json()
     assert response["success"]
-    assert len(response["result"]["moon"]) == automation.STORED_TRACES
+    assert len(response["result"]["moon"]) == automation.trace.STORED_TRACES
     assert len(response["result"]["sun"]) == 1
     assert int(response["result"]["moon"][0]["run_id"]) == int(moon_run_id) + 1
     assert (
         int(response["result"]["moon"][-1]["run_id"])
-        == int(moon_run_id) + automation.STORED_TRACES
+        == int(moon_run_id) + automation.trace.STORED_TRACES
     )
